refactor archivos
