+++ app/src/main/java/com/devs/agenda_agil/Fragment_backlog.java
@@ -10,6 +10,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.devs.entity.Tarea;
+
 import java.util.ArrayList;
