+++ app/src/main/java/com/devs/agenda_agil/Fragment_eventos.java
@@ -10,6 +10,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.devs.entity.Evento;
+
 import java.util.ArrayList;
