+++ app/src/main/java/com/devs/agenda_agil/TareasAdapter.java
@@ -1,12 +1,13 @@
 package com.devs.agenda_agil;
 
-import android.graphics.Color;
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import com.devs.entity.Tarea;
 
 import java.util.ArrayList;
