+++ app/src/main/java/com/devs/util/DateUtil.java
@@ -1,4 +1,4 @@
-package com.devs.src;
+package com.devs.util;
 
 import java.util.Calendar;
