+++ build.gradle
@@ -7,8 +7,10 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.4'
-        
+        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath "io.realm:realm-gradle-plugin:5.7.0"
+
+
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
